Fix typos
There is a typo in a provider name. While both `kubernetes` and
`hashicorp/kubernetes` are working option, I think it's better to specify that
Hashicord provider is required.

There is said in `main.tf` that `var.spec` can't be null. I don't see a reason
why it should be set to null by default. It's better when this parameter is
required and so I removed the default value for this variable.